Implement changes to get hw-python running for xApp Development Tutorial
This change fixes several bugs of hw-python xApp vanilla from OSC
to exchange messages with the e2-interface simulator vanilla also
from OSC.

Also, this change includes new features to deal with ASN1 messages
using the Pycrate library so that we can encode/deconde such messages
without the need to resort to C/C++ bindings.

Subscription also works out-of-the-box in this change.

Change-Id: I335a8fe55304d81bf59d94b500e91389b0446112